adding some design to transaction detail etc
